Commiting major changes to topological patroller the main change is due to the fact that the topological navigation server does not need origin node anymore.
